+++ example/test/example_test.mockor.dart
@@ -8,46 +8,46 @@ part of 'example_test.dart';
 // MockerGenerator
 // **************************************************************************
 
-@GenerateMocks([], customMocks: [
+@GenerateNiceMocks([
   MockSpec<ExampleUseCase>(
     as: #MockExampleUseCase,
-    returnNullOnMissingStub: true,
   ),
   MockSpec<ExampleUseCase2>(
     as: #MockExampleUseCase2,
-    returnNullOnMissingStub: true,
   ),
   MockSpec<ModelA.Model>(
     as: #MockModelAModel,
-    returnNullOnMissingStub: true,
   ),
   MockSpec<ModelB.Model>(
     as: #MockModelBModel,
-    returnNullOnMissingStub: true,
   ),
 ])
 dynamic _$_mock<T extends Object>({bool? relaxed}) {
   relaxed ??= false;
   switch (T) {
     case ExampleUseCase:
+    case MockExampleUseCase:
       final mock = MockExampleUseCase();
       if (!relaxed) {
         throwOnMissingStub(mock);
       }
       return mock;
     case ExampleUseCase2:
+    case MockExampleUseCase2:
       final mock = MockExampleUseCase2();
       if (!relaxed) {
         throwOnMissingStub(mock);
       }
       return mock;
     case ModelA.Model:
+    case MockModelAModel:
       final mock = MockModelAModel();
       if (!relaxed) {
         throwOnMissingStub(mock);
       }
       return mock;
     case ModelB.Model:
+    case MockModelBModel:
       final mock = MockModelBModel();
       if (!relaxed) {
         throwOnMissingStub(mock);